Teraz już wszystko działa. (Chyba)

+++ src/main/java/pl/javastart/minifilmweb/MovieRestController.java
@@ -1,6 +1,8 @@
 package pl.javastart.minifilmweb;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -12,6 +14,7 @@ import java.util.List;
 //        POST /api/movies/
 //        PUT /api/movies/{id}
 //        DELETE /api/movies/{id}
+
 @RestController
 @RequestMapping("/api/movies")
 public class MovieRestController {
@@ -26,8 +29,13 @@ public class MovieRestController {
     }
 
     @GetMapping("/{id}")
-    public Movie getMovie(@PathVariable("id") long id) {
-        return movieRepository.findOne(id);
+    @ResponseBody
+    public ResponseEntity<?> getMovie(@PathVariable("id") long id) {
+        Movie movie = movieRepository.findOne(id);
+        if (movie == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(movie);
     }
 
     @PostMapping("")
@@ -37,17 +45,27 @@ public class MovieRestController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteMovie(@PathVariable("id") long id) {
-        movieRepository.delete(id);
+    public ResponseEntity<?> deleteMovie(@PathVariable("id") long id) {
+        Movie movie = movieRepository.findOne(id);
+        if (movie == null) {
+            return new ResponseEntity<Movie>(HttpStatus.NOT_FOUND);
+        } else {
+            movieRepository.delete(id);
+        }
+        return ResponseEntity.ok(movie);
     }
 
     @PutMapping("/{id}")
-    public void editMovie(@RequestBody Movie movie,
-                          @PathVariable("id") long id) {
+    public ResponseEntity<?> editMovie(@RequestBody Movie movie,
+                                       @PathVariable("id") long id) {
         Movie movieEdit = movieRepository.findOne(id);
-        //movieEdit.setId(movie.getId());
+        if (movieEdit == null) {
+            return ResponseEntity.notFound().build();
+        }
         movieEdit.setTitle(movie.getTitle());
         movieEdit.setPremiereDate(movie.getPremiereDate());
         movieRepository.save(movieEdit);
+
+        return ResponseEntity.ok().build();
     }
 }
